Some things are better left unsigned

+++ src/test/kotlin/org/jbb/big/BigFileTest.kt
@@ -7,10 +7,10 @@ public class BigFileTest {
     Test fun testReadHeader() {
         BigBedFile.read(Examples.get("example1.bb")).use { bbf ->
             val header = bbf.header
-            assertEquals(1.toShort(), header.version)
-            assertEquals(5.toShort(), header.zoomLevelCount)
-            assertEquals(3.toShort(), header.fieldCount)
-            assertEquals(3.toShort(), header.definedFieldCount)
+            assertEquals(1, header.version)
+            assertEquals(5, header.zoomLevelCount)
+            assertEquals(3, header.fieldCount)
+            assertEquals(3, header.definedFieldCount)
             assertEquals(0, header.uncompressBufSize)
         }
     }
